Refactor: migrate Data Saver to Content Settings
